+++ docs/changes.rst
@@ -11,6 +11,7 @@ To be released. Still alpha version.
 - Added ``==`` operator for :class:`~wand.image.Image` type.
 - Added :func:`hash()` support of :class:`~wand.image.Image` type.
 - Added :attr:`Image.signature <wand.image.Image.signature>` property.
+- Changed the theme of Sphinx documentation.
 
 Version 0.1.8
 -------------